Build a new UI. Move the search query field out of the table. Move the handling of the filters to be in one place. Control the filters and reset the search query when changing the type file.
Disadvantages:
Lost the loading indicator of the table for (infinite-pagination, search for).
Have to implement our own debounce.
Have to pass the global filter on multiple levels, to the components and two of the API handlers custom hooks.
Have to move the GitHub table header to a standalone component since it has long jsx.
Have to make use query-params hook receive the filters as an object to use it once and let it handle all the logic.

Note:
To fix the table loader we just should set the enableTopToolbar property to true or remove it since its default value is true.